pkg/api: move *_test.go -> pkg/api/testing
