+++ pkg/api/testing/backward_compatibility_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package v1_test
+package testing
 
 import (
 	"testing"
