+++ pkg/api/testing/deep_copy_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package api_test
+package testing
 
 import (
 	"io/ioutil"
