+++ pkg/api/testing/node_example.json
