+++ pkg/api/testing/replication_controller_example.json
